% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_log_data.R
\name{clean_log_data}
\alias{clean_log_data}
\title{Clean raw log data frame}
\usage{
clean_log_data(data, drop_na = FALSE)
}
\arguments{
\item{data}{Raw log data frame}

\item{drop_na}{Drop rows containing missing values}
}
\value{
Clean log data frame
}
\description{
Clean raw log data frame
}
