% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_log.R
\name{read_raw_log}
\alias{read_raw_log}
\title{Read raw log file}
\usage{
read_raw_log(file, params_list = NULL, numeric_params = NULL)
}
\arguments{
\item{file}{Input log file}

\item{params_list}{Manual list of parameters (optional)}

\item{numeric_params}{List of numeric parameters (optional)}
}
\value{
Raw log data frame
}
\description{
Read raw log file
}
