% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_log_data.R
\name{summarise_log_data}
\alias{summarise_log_data}
\title{Summarise clean log data frame}
\usage{
summarise_log_data(data, params_list)
}
\arguments{
\item{data}{Clean log data frame}

\item{params_list}{Manual list of parameters}
}
\value{
Log summary data frame
}
\description{
Summarise clean log data frame
}
