% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_log_data.R
\name{summarise_log_data}
\alias{summarise_log_data}
\title{Summarise clean log data frame}
\usage{
summarise_log_data(
  data,
  params_list = NULL,
  num_folds = NULL,
  num_models = NULL
)
}
\arguments{
\item{data}{Clean log data frame}

\item{params_list}{Manual list of parameters (optional)}

\item{num_folds}{Number of folds per model/candidate (optional)}

\item{num_models}{Number of models/candidates (optional)}
}
\value{
Log summary data frame
}
\description{
Summarise clean log data frame
}
